Refactor game advance code to tell the caller to call it again.
The simulator logic was originally responsible for calling the state
handler logic when the CPU was supposed to do something. This will be
problematic because the renderer will eventually want to perform
animations when the CPU plays a card. This diff has the simulation code
tell the caller whether it should get user input or advance the game
state again, allowing for an opportunity to do other things in between
each game advancement.