Updates chose dealer logic to be explicit about human vs. cpu actions.
The state handling routines are getting confusing because the simulator
wants to do different things in the same state depending on if the human
is doing something or the CPU player should be doing something. This
breaks out the dealer cut state into two distinct states so we don't
confuse ourselves.